Moved config server related code to ConfigServerStartupHookProvider to make it pluggable feature
