+++ server/src/main/java/com/networknt/server/Server.java
@@ -16,9 +16,6 @@
 
 package com.networknt.server;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.networknt.client.Http2Client;
 import com.networknt.common.DecryptUtil;
 import com.networknt.common.SecretConstants;
 import com.networknt.config.Config;
@@ -36,43 +33,23 @@ import com.networknt.utility.Util;
 import io.undertow.Handlers;
 import io.undertow.Undertow;
 import io.undertow.UndertowOptions;
-import io.undertow.client.ClientConnection;
-import io.undertow.client.ClientRequest;
-import io.undertow.client.ClientResponse;
 import io.undertow.server.HttpHandler;
 import io.undertow.server.handlers.GracefulShutdownHandler;
 import io.undertow.util.Headers;
-import io.undertow.util.Methods;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.slf4j.MDC;
-import org.xnio.IoUtils;
-import org.xnio.OptionMap;
 import org.xnio.Options;
 import org.xnio.SslClientAuthMode;
 
 import javax.net.ssl.*;
-import java.io.BufferedInputStream;
-import java.io.FileOutputStream;
-import java.io.IOException;
 import java.io.InputStream;
 import java.net.InetAddress;
-import java.net.URI;
-import java.nio.file.FileSystem;
-import java.nio.file.FileSystems;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.security.KeyStore;
 import java.security.KeyStoreException;
 import java.security.NoSuchAlgorithmException;
 import java.security.UnrecoverableKeyException;
 import java.util.*;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipFile;
-
-import static java.nio.charset.StandardCharsets.UTF_8;
 
 /**
  * This is the entry point of the framework. It wrapped Undertow Core HTTP
@@ -85,27 +62,18 @@ public class Server {
 
     static final Logger logger = LoggerFactory.getLogger(Server.class);
     public static final String SERVER_CONFIG_NAME = "server";
-    public static final String STARTUP_CONFIG_NAME = "startup";
-    public static final String CONFIG_SECRET = "secret";
+    public static final String SECRET_CONFIG_NAME = "secret";
     public static final String[] STATUS_CONFIG_NAME = {"status", "app-status"};
-
-
+    
     public static final String ENV_PROPERTY_KEY = "environment";
     public static final String DEFAULT_ENV = "test";
 
     public static String lightEnv = System.getProperty("light-env");
-
-    static final String LIGHT_CONFIG_SERVER_URI = "light-config-server-uri";
-
-    static final String APIF_CONFIG_SERVER_URI = "apif-config-server-uri";
-    static final String APIF_CONFIG_SERVER_CONTEXT_ROOT = "/config-server/configs";
-
+    
     static final String STATUS_HOST_IP = "STATUS_HOST_IP";
 
     // service_id in slf4j MDC
     static final String SID = "sId";
-
-    public static StartupConfig startupConfig = (StartupConfig) Config.getInstance().getJsonObjectConfig(STARTUP_CONFIG_NAME, StartupConfig.class);
 
     public final static TrustManager[] TRUST_ALL_CERTS = new X509TrustManager[]{new DummyTrustManager()};
 
@@ -117,10 +85,7 @@ public class Server {
     static SSLContext sslContext;
 
     static GracefulShutdownHandler gracefulShutdownHandler;
-
-    // An instance of Jackson ObjectMapper that can be used anywhere else for Json.
-    final static ObjectMapper mapper = new ObjectMapper();
-
+    
     public static void main(final String[] args) {
         init();
     }
@@ -130,18 +95,18 @@ public class Server {
         // setup system property to redirect undertow logs to slf4j/logback.
         System.setProperty("org.jboss.logging.provider", "slf4j");
         // this will make sure that all log statement will have serviceId
-        MDC.put(SID, startupConfig.getServiceId());
+        MDC.put(SID, getServerConfig().getServiceId());
+
+        if (lightEnv == null) {
+            logger.warn("Warning! No light-env has been passed in from command line. Defaulting to {}",DEFAULT_ENV);
+            lightEnv = DEFAULT_ENV;
+        }
 
         try {
-            // load config properties from apif-config-server if possible.
-            loadConfigServerConfigs();
-
-            // load config files from light-config-server if possible.
-            //loadConfig();
+            start();
 
             // merge status.yml and app-status.yml if app-status.yml is provided
             mergeStatusConfig();
-            start();
         } catch (RuntimeException e) {
             // Handle any exception encountered during server start-up
             logger.error("Server is not operational! Failed with exception", e);
@@ -279,7 +244,7 @@ public class Server {
                 Map parameters = new HashMap<>();
                 if (lightEnv != null)
                     parameters.put(ENV_PROPERTY_KEY, lightEnv);
-                serviceUrl = new URLImpl("light", ipAddress, port, startupConfig.getServiceId(), parameters);
+                serviceUrl = new URLImpl("light", ipAddress, port, getServerConfig().getServiceId(), parameters);
                 registry.register(serviceUrl);
                 if (logger.isInfoEnabled())
                     logger.info("register service: " + serviceUrl.toFullStr());
@@ -370,7 +335,7 @@ public class Server {
 
     private static KeyStore loadKeyStore() {
         ServerConfig serverConfig = getServerConfig();
-        Map<String, Object> secretConfig = DecryptUtil.decryptMap(Config.getInstance().getJsonMapConfig(CONFIG_SECRET));
+        Map<String, Object> secretConfig = DecryptUtil.decryptMap(Config.getInstance().getJsonMapConfig(SECRET_CONFIG_NAME));
 
         String name = serverConfig.getKeystoreName();
         try (InputStream stream = Config.getInstance().getInputStreamFromFile(name)) {
@@ -385,7 +350,7 @@ public class Server {
 
     protected static KeyStore loadTrustStore() {
         ServerConfig serverConfig = getServerConfig();
-        Map<String, Object> secretConfig = DecryptUtil.decryptMap(Config.getInstance().getJsonMapConfig(CONFIG_SECRET));
+        Map<String, Object> secretConfig = DecryptUtil.decryptMap(Config.getInstance().getJsonMapConfig(SECRET_CONFIG_NAME));
 
         String name = serverConfig.getTruststoreName();
         try (InputStream stream = Config.getInstance().getInputStreamFromFile(name)) {
@@ -433,7 +398,7 @@ public class Server {
 
     private static SSLContext createSSLContext() throws RuntimeException {
         ServerConfig serverConfig = getServerConfig();
-        Map<String, Object> secretConfig = DecryptUtil.decryptMap(Config.getInstance().getJsonMapConfig(CONFIG_SECRET));
+        Map<String, Object> secretConfig = DecryptUtil.decryptMap(Config.getInstance().getJsonMapConfig(SECRET_CONFIG_NAME));
 
         try {
             KeyManager[] keyManagers = buildKeyManagers(loadKeyStore(),
@@ -455,188 +420,6 @@ public class Server {
         }
     }
 
-    private static void loadConfigServerConfigs(){
-        if (lightEnv == null) {
-            logger.warn("Warning! No light-env has been passed in from command line. Defaulting to {}",DEFAULT_ENV);
-            lightEnv = DEFAULT_ENV;
-        }
-        String configUri = System.getProperty(APIF_CONFIG_SERVER_URI);
-        if (configUri != null) {
-            StringBuilder configPath = new StringBuilder();
-            configPath.append("/").append(startupConfig.getProjectId());
-            configPath.append("/").append(startupConfig.getProjectVersion());
-            configPath.append("/").append(startupConfig.getServiceId());
-            configPath.append("/").append(startupConfig.getServiceVersion());
-            configPath.append("/").append(lightEnv);
-            logger.debug("configPath: {}", configPath);
-
-            //get service configs and put them in config cache
-            Map<String, Object> serviceConfigs = getServiceConfigs(configPath.toString());
-            //set the environment value (the one used to fetch configs) in the serviceConfigs going into configCache
-            serviceConfigs.put(ENV_PROPERTY_KEY,lightEnv);
-            logger.debug("serviceConfigs: {}", serviceConfigs);
-            //clear config cache: this is required just in case other classes have already loaded something in cache
-            Config.getInstance().clear();
-            Config.getInstance().putInConfigCache(Config.CENTRALIZED_MANAGEMENT, serviceConfigs);
-        }else {
-            logger.info("Property {} is missing in the command line. Using local config files",APIF_CONFIG_SERVER_URI);
-        }
-    }
-
-    private static Map<String, Object> getServiceConfigs(String configPath) {
-        String authorization=System.getenv("Authorization");
-        Map<String, Object> configs = new HashMap<>();
-        String configServerHost = System.getProperty(APIF_CONFIG_SERVER_URI);
-        String configServerPath =  APIF_CONFIG_SERVER_CONTEXT_ROOT + configPath;
-
-        logger.debug("Calling Config Server endpoint:{}{}", configServerHost, configServerPath);
-
-        Http2Client client = Http2Client.getInstance();
-        ClientConnection connection = null;
-        try {
-            connection = client.connect(new URI(configServerHost), Http2Client.WORKER, Http2Client.SSL, Http2Client.BUFFER_POOL,
-                    OptionMap.create(UndertowOptions.ENABLE_HTTP2, true)).get();
-
-            final CountDownLatch latch = new CountDownLatch(1);
-            final AtomicReference<ClientResponse> reference = new AtomicReference<>();
-
-            ClientRequest request = new ClientRequest().setMethod(Methods.GET).setPath(configServerPath);
-            request.getRequestHeaders().put(Headers.HOST, configServerHost);
-            request.getRequestHeaders().put(Headers.AUTHORIZATION, authorization);
-            connection.sendRequest(request, client.createClientCallback(reference, latch));
-            latch.await();
-
-            int statusCode = reference.get().getResponseCode();
-
-            if (statusCode >= 300) {
-                logger.error("Failed to load configs from config server" + statusCode + ":"
-                        + reference.get().getAttachment(Http2Client.RESPONSE_BODY));
-                throw new Exception("Failed to load configs from config server: " + statusCode);
-            } else {
-                String respBody = reference.get().getAttachment(Http2Client.RESPONSE_BODY);
-                Map<String, Object> response = (Map<String, Object>) mapper.readValue(respBody, new TypeReference<Map<String, Object>>() {});
-                configs = (Map<String, Object>) response.get("configProperties");
-            }
-        } catch (Exception e) {
-            logger.error("Exception while calling config server:", e);
-        } finally {
-            IoUtils.safeClose(connection);
-        }
-
-        return configs;
-    }
-
-    /**
-     * Load config files from light-config-server instance. This is normally only
-     * used when you run light-4j server as standalone java process. If the server
-     * is dockerized and orchestrated by Kubernetes, the config files and secret
-     * will be mapped to Kubernetes ConfigMap and Secret and passed into the
-     * container.
-     * <p>
-     * Of course, you can still use it with standalone docker container but it is
-     * not recommended.
-     */
-    private static void loadConfig() {
-        // if it is necessary to load config files from config server
-        // Here we expect at least env(dev/sit/uat/prod) and optional config server url
-        if (lightEnv == null) {
-            logger.warn("Warning! No light-env has been passed in from command line. Default to dev");
-            lightEnv = DEFAULT_ENV;
-        }
-        String configUri = System.getProperty(LIGHT_CONFIG_SERVER_URI);
-        if (configUri != null) {
-            // try to get config files from the server.
-            String targetMergeDirectory = System.getProperty(Config.LIGHT_4J_CONFIG_DIR);
-            if (targetMergeDirectory == null) {
-                logger.warn("Warning! No light-4j-config-dir has been passed in from command line.");
-                return;
-            }
-            String version = Util.getJarVersion();
-            String service = startupConfig.getServiceId();
-            String tempDir = System.getProperty("java.io.tmpdir");
-            String zipFile = tempDir + "/config.zip";
-            // /v1/config/1.2.4/dev/com.networknt.petstore-1.0.0
-
-            String path = "/v1/config/" + version + "/" + lightEnv + "/" + service;
-            Http2Client client = Http2Client.getInstance();
-            ClientConnection connection = null;
-            try {
-                connection = client.connect(new URI(configUri), Http2Client.WORKER, Http2Client.SSL, Http2Client.BUFFER_POOL,
-                        OptionMap.create(UndertowOptions.ENABLE_HTTP2, true)).get();
-            } catch (Exception e) {
-                logger.error("Exeption:", e);
-            }
-            final CountDownLatch latch = new CountDownLatch(1);
-            final AtomicReference<ClientResponse> reference = new AtomicReference<>();
-            try {
-                ClientRequest request = new ClientRequest().setMethod(Methods.GET).setPath(path);
-                request.getRequestHeaders().put(Headers.HOST, "localhost");
-                connection.sendRequest(request, client.createClientCallback(reference, latch));
-                latch.await();
-                int statusCode = reference.get().getResponseCode();
-
-                if (statusCode >= 300) {
-                    logger.error("Failed to load config from config server" + statusCode + ":"
-                            + reference.get().getAttachment(Http2Client.RESPONSE_BODY));
-                    throw new Exception("Failed to load config from config server: " + statusCode);
-                } else {
-                    // TODO test it out
-                    FileOutputStream fos = new FileOutputStream(zipFile);
-                    fos.write(reference.get().getAttachment(Http2Client.RESPONSE_BODY).getBytes(UTF_8));
-                    fos.close();
-                    unzipFile(zipFile, targetMergeDirectory);
-                }
-            } catch (Exception e) {
-                logger.error("Exception:", e);
-            } finally {
-                IoUtils.safeClose(connection);
-            }
-        } else {
-            logger.info("light-config-server-uri is missing in the command line. Use local config files");
-        }
-    }
-
-    private static void mergeConfigFiles(String source, String target) {
-
-    }
-
-    private static void unzipFile(String path, String target) {
-        // Open the file
-        try (ZipFile file = new ZipFile(path)) {
-            FileSystem fileSystem = FileSystems.getDefault();
-            // Get file entries
-            Enumeration<? extends ZipEntry> entries = file.entries();
-
-            // We will unzip files in this folder
-            Files.createDirectory(fileSystem.getPath(target));
-
-            // Iterate over entries
-            while (entries.hasMoreElements()) {
-                ZipEntry entry = entries.nextElement();
-                // If directory then create a new directory in uncompressed folder
-                if (entry.isDirectory()) {
-                    System.out.println("Creating Directory:" + target + entry.getName());
-                    Files.createDirectories(fileSystem.getPath(target + entry.getName()));
-                }
-                // Else create the file
-                else {
-                    InputStream is = file.getInputStream(entry);
-                    BufferedInputStream bis = new BufferedInputStream(is);
-                    String uncompressedFileName = target + entry.getName();
-                    Path uncompressedFilePath = fileSystem.getPath(uncompressedFileName);
-                    Files.createFile(uncompressedFilePath);
-                    FileOutputStream fileOutput = new FileOutputStream(uncompressedFileName);
-                    while (bis.available() > 0) {
-                        fileOutput.write(bis.read());
-                    }
-                    fileOutput.close();
-                    System.out.println("Written :" + entry.getName());
-                }
-            }
-        } catch (IOException e) {
-            logger.error("IOException", e);
-        }
-    }
 
     // method used to merge status.yml and app-status.yml
     protected static void mergeStatusConfig() {
