--- server/src/main/java/com/networknt/server/StartupConfig.java
@@ -1,41 +1,0 @@
-package com.networknt.server;
-
-public class StartupConfig {
-    public static final String CONFIG_NAME = "startup";
-    private String projectId;
-    private String projectVersion;
-    private String serviceId;
-    private String serviceVersion;
-
-    public String getProjectId() {
-        return projectId;
-    }
-
-    public void setProjectId(String projectId) {
-        this.projectId = projectId;
-    }
-
-    public String getProjectVersion() {
-        return projectVersion;
-    }
-
-    public void setProjectVersion(String projectVersion) {
-        this.projectVersion = projectVersion;
-    }
-
-    public String getServiceId() {
-        return serviceId;
-    }
-
-    public void setServiceId(String serviceId) {
-        this.serviceId = serviceId;
-    }
-
-    public String getServiceVersion() {
-        return serviceVersion;
-    }
-
-    public void setServiceVersion(String serviceVersion) {
-        this.serviceVersion = serviceVersion;
-    }
-}
